+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/CPDConfiguration.java
@@ -67,6 +67,9 @@ public class CPDConfiguration extends AbstractConfiguration {
     @Parameter(names = "--ignore-annotations", description = "Ignore language annotations when comparing text", required = false)
     private boolean ignoreAnnotations;
 
+    @Parameter(names = "--ignore-usings", description = "Ignore using directives in C#", required = false)
+    private boolean ignoreUsings;
+
     @Parameter(names = "--skip-lexical-errors", description = "Skip files which can't be tokenized due to invalid characters instead of aborting CPD", required = false)
     private boolean skipLexicalErrors = false;
 
@@ -227,6 +230,11 @@ public class CPDConfiguration extends AbstractConfiguration {
         } else {
             properties.remove(Tokenizer.IGNORE_ANNOTATIONS);
         }
+        if (configuration.isIgnoreUsings()) {
+            properties.setProperty(Tokenizer.IGNORE_USINGS, "true");
+        } else {
+            properties.remove(Tokenizer.IGNORE_USINGS);
+        }
         properties.setProperty(Tokenizer.OPTION_SKIP_BLOCKS, Boolean.toString(!configuration.isNoSkipBlocks()));
         properties.setProperty(Tokenizer.OPTION_SKIP_BLOCKS_PATTERN, configuration.getSkipBlocksPattern());
         configuration.getLanguage().setProperties(properties);
@@ -336,6 +344,14 @@ public class CPDConfiguration extends AbstractConfiguration {
 
     public void setIgnoreAnnotations(boolean ignoreAnnotations) {
         this.ignoreAnnotations = ignoreAnnotations;
+    }
+
+    public boolean isIgnoreUsings() {
+        return ignoreUsings;
+    }
+
+    public void setIgnoreUsings(boolean ignoreUsings) {
+        this.ignoreUsings = ignoreUsings;
     }
 
     public boolean isSkipLexicalErrors() {