+++ pmd-cs/src/main/java/net/sourceforge/pmd/cpd/CsLanguage.java
@@ -3,15 +3,24 @@
  */
 package net.sourceforge.pmd.cpd;
 
+import java.util.Properties;
+
 /**
  * Language implementation for C#
  */
 public class CsLanguage extends AbstractLanguage {
 
-    /**
-     * Creates a new C# Language instance.
-     */
     public CsLanguage() {
+        this(System.getProperties());
+    }
+
+    public CsLanguage(Properties properties) {
         super("C#", "cs", new CsTokenizer(), ".cs");
+        setProperties(properties);
+    }
+
+    public final void setProperties(Properties properties) {
+        CsTokenizer tokenizer = (CsTokenizer)getTokenizer();
+        tokenizer.setProperties(properties);
     }
 }