+++ src/site/markdown/overview/changelog.md
@@ -6,10 +6,13 @@
 
 **Feature Request and Improvements:**
 
+*   CPD: New command line parameter `--ignore-usings`: Ignore using directives in C# when comparing text.
+
 **New/Modified/Deprecated Rules:**
 
 **Pull Requests:**
 
+*   [#25](https://github.com/adangel/pmd/pull/25): Added option to exclude C# using directives from CPD analysis
 *   [#72](https://github.com/pmd/pmd/pull/72): Added capability in Java and JSP parser for tracking tokens.
 *   [#74](https://github.com/pmd/pmd/pull/74): Fix rendering CommentDefaultAccessModifier description as code
 *   [#75](https://github.com/pmd/pmd/pull/75): RuleSetFactory Performance Enhancement